added connection from scrabble to lib folder. passed first test for scoring word
