+++ lib/scoring.rb
@@ -1,0 +1,28 @@
+class Scoring
+  #some sort of data structure to store the individual letter scores
+  SCORE_CHART = {
+    "A"=>1, "B"=>3, "C"=>3, "D"=>2,
+    "E"=>1, "F"=>4, "G"=>2, "H"=>4,
+    "I"=>1, "J"=>8, "K"=>5, "L"=>1,
+    "M"=>3, "N"=>1, "O"=>1, "P"=>3,
+    "Q"=>10, "R"=>1, "S"=>1, "T"=>1,
+    "U"=>1, "V"=>4, "W"=>4, "X"=>8,
+    "Y"=>4, "Z"=>10
+  }
+  def self.score(word)
+    word_array = word.upcase.split("")
+
+    word_score = 0
+    raise ArgumentError unless word.length < 8
+
+    word_array.each do |letter|
+      word_score += SCORE_CHART[letter]
+    #returns score value for given word. word input as string.
+    end
+    #seven letter word get 50point bonus
+    return word_score
+  end
+
+  def self.highest_score_from(array_of_words)
+  end
+end
